requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
cssselect2==0.7.0
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.6
fonttools==4.55.3
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
openai==1.58.1
packaging==24.2
pillow==11.1.0
portalocker==2.10.1
postgrest==0.18.0
propcache==0.2.1
protobuf==5.29.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydyf==0.11.0
pyphen==0.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
qdrant-client==1.12.2
realtime==2.0.6
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==63.1
webencodings==0.5.1
websockets==13.1
yarl==1.18.3
zopfli==0.2.3.post1
